Throw exception instead of printing error message and terminating the script
In case of an error with Fomo’s API it was impossible to handle the
situation gracefully as the application was terminated by calling
`exit()`. Printing an error message to a potentially user-facing
client is a bad idea, too. This commit changes that behaviour by
throwing exceptions which can be catched by the caller.